py3/ta/requirements.txt
backports-abc
beautifulsoup4
bs4
future
markdown2
pyparsing
python-dateutil
PyYAML
singledispatch
six
tornado
gitpython

